working on drag and drop
