+++ src/TodoList.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import TodoListItem from './TodoListItem'
-import TaskEdit from './TaskEdit'
+// import TaskEdit from './TaskEdit'
 
 const placeholder = document.createElement("li");
 placeholder.className = "placeholder";
